testing/web-platform/tests/webmessaging/without-ports/008.html

<!doctype html>
<title>just one argument</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws('NOT_SUPPORTED_ERR', function() {
    postMessage('');
  });
});
</script>
